feat: update config and schedule (#106)
