+++ __tests__/i.test.ts
@@ -1,4 +1,5 @@
 import { expect, test } from 'vitest'
+
 const sum = (a: number, b: number): number => a + b
 
 test('adds 1 + 2 to equal 3', () => {
