+++ package.json
@@ -17,11 +17,11 @@
     "checkAll": "biome check . --apply-unsafe"
   },
   "dependencies": {
-    "@vkontakte/icons": "2.90.0",
-    "@vkontakte/vk-bridge": "^2.12.1",
+    "@vkontakte/icons": "2.95.0",
+    "@vkontakte/vk-bridge": "^2.13.0",
     "@vkontakte/vk-bridge-react": "^1.0.0",
-    "@vkontakte/vk-mini-apps-router": "^1.2.0",
-    "@vkontakte/vkui": "^6.0.0-beta.3",
+    "@vkontakte/vk-mini-apps-router": "^1.4.2",
+    "@vkontakte/vkui": "6.0.0",
     "jshashes": "^1.0.8",
     "preact": "^10.19.2",
     "react": "npm:@preact/compat",
@@ -29,19 +29,19 @@
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.23.3",
-    "@biomejs/biome": "1.4.1",
-    "@diary-spo/shared": "1.8.8",
+    "@biomejs/biome": "1.5.2",
+    "@diary-spo/shared": "1.8.9",
     "@diary-spo/types": "latest",
     "@preact/preset-vite": "^2.8.1",
     "@rollup/plugin-terser": "^0.4.4",
-    "@types/node": "^20.10.7",
-    "@vitest/coverage-v8": "^1.1.3",
+    "@types/node": "^20.11.5",
+    "@vitest/coverage-v8": "^1.2.1",
     "@vkontakte/vk-miniapps-deploy": "^0.1.6",
     "gh-pages": "^6.0.0",
     "jsdom": "^23.2.0",
     "typescript": "^5.2.2",
-    "vite": "^5.0.11",
-    "vitest": "^1.2.0",
+    "vite": "^5.0.12",
+    "vitest": "^1.2.1",
     "vitest-localstorage-mock": "^0.0.1"
   }
 }
