+++ src/config/index.ts
@@ -1,5 +1,3 @@
-export const BASE_URL = import.meta.env.VITE_SERVER_URL
-export const SECOND_SERVER_URL = import.meta.env.VITE_SERVER_URL_SECOND
-
 export * from './colors.ts'
 export * from './constants.ts'
+export * from './env.ts'
