+++ src/methods/makeRequest.ts
@@ -1,5 +1,5 @@
 import { BASE_URL } from '@config'
-import { ServerResponse } from '../types'
+import { HTTP_STATUSES, ServerResponse } from '../types'
 import requestToSecondServer from './requestToSecondServer.ts'
 
 const makeRequest = async <T>(
@@ -26,13 +26,22 @@ const makeRequest = async <T>(
 
     clearTimeout(timeoutId)
 
+    /** В случае ошибки авторизации мы не делаем запрос на второй сервер, а сразу возвращаем ответ **/
+    if (response.status === HTTP_STATUSES.UNAUTHORIZED) {
+      return response
+    }
+
+    console.info('%c [makeRequest]', 'color: blueviolet', response)
+
+    /** В случае другой ошибки пытаемся получить ответ от второго сервера **/
     if (!response.ok) {
       return requestToSecondServer(route, token, method, body)
     }
 
     return (await response.json()) as T
   } catch (err) {
-    console.error(err)
+    console.info('%c [makeRequest]', 'color: blueviolet', err)
+    /** В случае ошибки пытаемся получить ответ от второго сервера **/
     return requestToSecondServer(route, token, method, body)
   }
 }
