+++ src/views/Contacts/data.ts
@@ -1,23 +1,14 @@
-import { IHelpAccordion } from './HelpAccordion'
-
-export const helpData: IHelpAccordion[] = [
-  {
-    id: 1,
-    title: 'Часто приходится удалять сервис',
-    detail:
-      'Как правило, эта проблема появляется на телефонах. Одно из возможных решений — не заходить по прямой ссылке в ' +
-      'сервис, а через страницу Сервисы -> Для вас -> Дневник СПО.'
-  },
+export const helpData = [
   {
     id: 2,
-    title: 'В сервисе стоит оценка, которой нет в журнале Сетевого города',
+    title: 'Не совпадают оценки',
     detail:
       "Иногда в журнале может появиться отметка, которой нет в оригинальном дневнике (Сетевой город). Обычно это 'Д'" +
       ' (Долг), но если вы уверены, что долга у вас нет, то напишите нам.'
   },
   {
     id: 3,
-    title: 'На странице Успеваемость нет данных обо мне',
+    title: 'Неправильные данные',
     detail:
       'Если на странице Успеваемость нет данных о вас, то перезайдите в сервис (Настройки - Выйти), но если они ' +
       'неправильные, то сообщите нам. Все данные, кроме аватарки, мы берём из Сетевого города.'
@@ -28,6 +19,6 @@ export const helpData: IHelpAccordion[] = [
     detail:
       'Если вы при заходе на любую страницу получаете ошибку загрузки данных с сервера, то тут может быть несколько ' +
       'проблем: 1. Сервер сетевого города или наш упал(и) — обычно мы сообщаем об этом в группе. 2. Ваша cookie ' +
-      'устарела и вам надо получить новую — перезайдите в аккаунт или обновите её в настройках.'
+      'устарела и вам надо получить новую — перезайдите в аккаунт и сообщите нам об этом.'
   }
 ]
